Test coverage of sourcemap improvements
