+++ package.json
@@ -36,7 +36,7 @@
     "publish": "vsce publish",
     "publish-pre-release": "vsce publish --pre-release",
     "vscode:prepublish": "rimraf dist && npm run -S esbuild-base -- --minify",
-    "test": "ts-mocha -p tsconfig.json tests/*test.ts --timeout 30s"
+    "test": "ts-mocha -p tsconfig.json tests/*test.ts --timeout 30s --diff false"
   },
   "devDependencies": {
     "@types/glob": "^7.2.0",
