+++ sampleWorkspace/sourcemap-test/lib.teal
@@ -1,0 +1,2 @@
+#pragma version 9
+#define assertEquals ==; assert
